modificaciones calase Event y EventCard

+++ app/src/main/java/com/example/willgo/data/Event.kt
@@ -14,7 +14,7 @@ data class Event(
     val date: String? = "",
     val price: Float?,
     val image: String?,
-    val duration: Long?,
+    val duration: Float?,
     val asistance: Long?,
     val type: String? = ""
 )
